pyproject.toml
[project]
name = "spotifyPlaylistBackups"
version = "0.1.0"
description = "Backup and export Spotify playlists to JSON"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dropbox>=11.36.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.1",
    "PyYAML>=6.0",
    "python-dotenv>=1.0.0",
    "spotipy>=2.23.0",
    "typer[all]>=0.12.3",
]

[tool.black]
line-length = 127
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | venv
  | build
  | dist
  | __pycache__
)/
'''

[tool.isort]
profile = "black"
line_length = 127
skip_gitignore = true
skip = [".venv", "venv", "build", "dist"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--strict-markers",
    "--disable-warnings",
]

[tool.mypy]
python_version = "3.10"
strict = true
warn_unused_configs = true
ignore_missing_imports = true
explicit_package_bases = true
files = ["src"]
exclude = [
    'venv',
    '.venv',
    'build',
    'dist',
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/venv/*",
    "*/.venv/*",
    # Add whitelisted files here - files excluded from coverage requirements
    # Example: Interactive CLI scripts, cloud provider implementations
]

[tool.coverage.report]
# Note: fail_under is NOT set here - it's only enforced in the dedicated
# coverage CI job via --cov-fail-under=95 to keep it non-voting
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    # Whitelisted patterns - excluded from coverage requirements
    "def main\\(\\):",  # Main entry points (often hard to unit test)
    "@abstractmethod",  # Abstract methods
]
